Refactor flake overlays so they can be composed
Using "override" and "packageOverrides" like

    lua = prev.lua.override {
        packageOverrides = this: other: {
            enum = lua_enum;
        };
    };

does not make the overlay composable.

That is, if used like

    pkgs = import <nixpkgs> {
        overlays = [
            enum-overlay
            other-lua-overlay
        ];
    };

The *last* overlay cancels out the packageOverrides of the previous one. That
is, in the above example, there would be no "luaPackages.enum" attribute.

Using something like

      luaPackages = prev.luaPackages // { enum = lua_enum; };

instead, seems to be working.

Also: fix a bug where lua52_enum as passed to lua5_1.